backend/requirements.txt
flask>=2.2
python-dotenv>=1.0
psycopg2-binary>=2.9
qdrant-client>=1.7
bcrypt>=4.0
langchain>=0.1.0
sentence-transformers>=2.2
requests>=2.28
langchain-google-genai>=0.0.8
google-generativeai>=0.8.3
flask-cors
pandas>=2.0
minio>=7.1
openpyxl>=3.1
xlrd>=2.0
numpy>=1.24
scikit-learn>=1.3
tiktoken>=0.5